AMR information panel fix
The number of cells and the number of points for AMR data was wrong in
PV. This is because AMR do not index the internal data sets in the same
way other composite data sets do.
Extracting the data sets on which to display information using
`vtkExtractBlockUsingDataAssembly` should always be used on such data
sets, which is what this commit ensures.

Fixes #20947