FindData: Invert Selection bug fix
When selecting cells or points using `Select Blocks` or equivalents in
the selection toolbar, the Find Data widget was unable to invert the
selection, like it was in 5.9.1.

In order to be able to invert any input that was given by the user, the
invert checkbox needed to be moved in the display panel, with run-time
handling of the invert flag directly written in
`pqFindDataCurrentSelectionFrame`.

After this commit, `FindData` widget behaves the same was as its
previous dialog form w.r.t. inverting selections.