Fix problems with mixed element grids
Because the CGNS file groups cells of the same type in different sections
in a zone, the order of the cells in the file can be different than the
order in the cell-data arrays. To compensate for this, when writing cell-
centered data, the data is reordered to the order that the cells have in the
CGNS file.

(cherry picked from commit 376c44bdaf5f62710ea8668a4865d82f24a33545)