pqTransferFunctionWidget: ensure interaction ends
When a node in the pqTransferFunctionWidget is double-clicked, a color
chooser dialog appears to change the color. When the dialog is closed,
if the mouse is moved over the transfer function widget right away, a
"mouse button released" event can be lost, which results in the editor
getting stuck in a mode where the node moves based on the movement of
the mouse, resulting in an unintentional move of the node.

Fix this by creating a mouse button release event and sending it to
the selected vtkControlPointsItem to let it know interaction is
finished after the color editor dialog is closed.