Link the 2D transfer function properties
This change links the properties exposed on the properties panel with
the color opacity editor properties for 2D transfer functions.
It also ensures that the histogram in the 2D transfer function editor of
the color opacity editor panel is outdated when properties like
useGradientForTransfer2D or ColorArray2Name are changed.