DisplaySizedImplicitPlaneWidget: Pick Origin and Normal
`vtkPVRayCastPickingHelper` is now responsible for calculating both the intersection point and the normal.

`vtkSMRenderViewProxy` now actually returns if picking a point on the surface was successful or not.

`pqPointPickingHelper` can now pick a point or a normal and can enable or not (default is disabled) the picking of the camera focal point/normal.