Added *.encas file extension for Ensight Case Gold reader.
ANSYS Fluent exports Ensight Case Gold files with the *.encas file extension, requiring renaming of the extension to *.case for Paraview to recognize the file.  For a single file this is trival, but for transients data or other instances where there are many files, it becomes non-trivial and error prone to manage.

The extension problem is mentioned (but is not the primary purpose of) issue #14611.