Change file name in reader
After this commit, the user can change the file name of a
reader that
is already part of a pipeline. The file name in the
pipeline browser
changes accordingly, and the selection state of
arrays from the new array selection that were
present in the previous file are kept.

Fixes #21124