pqFileDialog: "Look in" not synchronised with model bug fix
The addition of `groupFiles` parameter (see #18000) had the consequence
to unlink what is written in "Look in" in the file dialog from the
directory being displayed. This was because of the default parameter for
`groupFiles` (set to true) which was causing an ambiguity on the version
of the method `pqImplementation::GetData` to use in
`vtkFileDialogModel`. Removing the default value and manually putting
`true` everywhere fixes the issue.

Fixes #21139