New option in the scalar bar to show displayed range values
The range values of the visible scalar fields can now be displayed on
the scalar bar.
To do so, click on `Edit Color Map` icon, then in the new menu, click on
`Edit color legend properties` icon. If the checkbox `Show Range` is on,
then the range will be
printed next to the color bar.

The range printed is composed of a Min value, and a Max value. The Min
value (resp. Max value)  is the minimum (resp. maxmimum) value of
all combined arrays that share the same name and are visible in the view
(with an open eyeball in the pipeline browser).

The range values are updated each time
`vtkSMPVRepresentationProxy::ViewUpdated` is called.
To handle scalar bar separation, special treatment is needed in
`pqUseSeparateColorMapReaction`

`vtkContext2DScalarBarActor` needed
enhanced with new attributes `DrawDataRange` and `MinRange` and
`MaxRange`

Fixes #21122