New ReplaceReaderFileName in simple.py
`ReplaceReaderFileName` is a new function in `simple.py` that aims at
hiding the backend of the `Change File` option. This way,
if this functionality needs amended, we can do so without breaking
already existing python traces.

When tracing with python, the new
`vtkSMCoreUtilities::ReplaceReaderFileName` is explicitly called.

`vtkSMTrace` can now trace functions that need a list of strings as a
parameter.