feat(OpenXR): add support for oculus touch controllers
Add support for oculus touch controllers in the OpenVR plugin
when running under the OpenXR runtime.

Fix a couple errors in the khr simple controller json file
as well.

Note the bindings for triggers are not supported yet so
next view is the menu button on the left controller and
the right trigger action uses button A for now.