Scalar Bar Range display bug fix
When changing the array to be displayed on a representation that was
sharing a scalar bar with another representation, the previous array's
range did not get properly updated. This was because the proxy was still
stored in the scalar bar widget when it should have been removed. The
proxy is now removed at update of scalar bar visibility when the lut
property changed.

Fixes #21122