Fixing GMVReaderOneOrTwoVertices test failure on ARM MacOs
ARM MacOs currently fail to render lines. The test
`GMVReaderOneOrTwoVertices` was using lines to test the output geometry
of the `GMVReader` on inputs that have vertices and lines as cells. This
test is removed and replaced by a python test that doesn't require
rendering. The number of vertices and lines in the output is verified.