Added support for multiple 2d levels in one file.
CDI can assign different vertical axis id to 2d fields depending on
netcdf atrribues of the variable (level_type=abc).
We better accept that they all are 2d and merge them into one 2D thing.

working towards support for multiple grids -- problems w/ ext grids.

External grid files currently are trouble.

Basically, the grid ids are diffent in the files, and we need to keep
track of that and of which file we are looking at.

Will be solved by having the grid and the data open simultaneously in
commits to come.

Starting to move file related things outside the main reader.

Fixups for grid handling

fixed a segfault in parallel mode on large grids

double clon_vert[size2] is illegal in c++, as size2 is not known at
compile time. On DKRZ's levante it leads to a segfault when accessing this
for large (5km global) grids in parallel operation.

Do not use BuildVarArrays for checking whether the variables are valid

BuildVarArrays enables all variables in the SelectionStatus

Caught a SetDimensions in statefile loading

SetDimensinos gets called before setting the file name when loading a
state file. This gets rid of a warning message in that case.