Add SpaceMouseInteractor plugin, controlling the camera with a SpaceMouse
Using the 3DxWareSDK provided by 3Dconnexion, connect to a SpaceMouse
if it is available, and use it to manipulate the camera in the active
renderview. The 3Dconnexion Settings app provided with the driver
can be used to change the style of interaction from object-centered
to camera-centered (flying).

Only works on Windows currently and provides an empty placeholder
for linux/mac. A Mac implementation is planned as a follow-up.
3Dconnexion does not provide an equivalent SDK for linux.