Insade 2D - 3D transition
When going from 2D view to 3D view on a 2D poly data (on the Mandelbrot
source for instance), when switching back to 3D, the object is very very
far, and the user needs to zoom in. The camera's z coordinate was being
set to 10000. This commit sets it to 3.35 * maxWidth, where maxWidth is
the max width of the bounding box.
This returns the same default value as in views_and_representations.xml
for bounds [-10, 10, -10, 10, -10, 10].

Fixes #21300