pqTextLocationWidget: update coordinates when switching to AnyLocation
For chart representations that use this widget, changing the window
location to fixed locations in the view should change the Position
coordinates in the VTK object. When switching to the AnyLocation mode,
update the Position property with the value by retrieving the
PositionInfo property from the server if that property exists.