Initializing the scalar bar when volume rendering
Volume rendering implies setting up a scalar field to render. If nothing
was selected, the first one in the scalar field list is taken by
default. However the scalar bar linked to the representation and the
view was not initialized.

The scalar bar mechanism in `pqDisplayColorWidget` has been turned into
a static method that is accessed by `pqDisplayRepresentationWidget`
when trying to set up the `Volume` representation.

Fixes #21452