Updating volume rendering tests to account for automatic scalar bar
The scalar bar behavior was fixed for volume rendering (it was not
turning on when displaying scalars). This caused several tests to fail
because they didn't assume a scalar bar being rendered. The scalar bar
is turned off in all those tests.

The test ImageVolumeRendering is untouched, but the baseline is updated.
This will ensure that the scalar bar behavior stays correct in the
future.