Add a Misc setting that limits the types shown in "File .. Open"
In the settings dialog, using a custom widget, show a list of all known
readers, and checkboxes to choosed whether a reader is shown in "File ..
Open". Just the excluded readers are stored, so any new readers are enabled
by default.

`pqLoadDataReaction` retrieves this list and skips readers that match.