Handle tilde (~) on Linux when loading state files
Not sure why but QFileInfo doesn't seem to handle tilde (~)
on Linux machines. It doesn't seem to be considered either
an absolute or relative path so if tilde is used for specifying
the state file location on the command line QFileInfo says that
the file isn't found. This fixes that issue.

Fix #21470