vtkPVClipDataSet: Optimize Plane Clips when ExactBoxClip is on
When ExactBoxClip is on, the ordering of the clips matter performance
wise, because the smaller clip output you create in the i-th clip the
better performance will be in the i+1-th clip. To achieve a good
ordering, we clip the bounding box and calculate the volume of the result.
The plane that generates the smallest volume goes first. This is done
recursively. Also planes that don't intersect with the bounding box are
removed if they don't point towards the dataset.