Sci Viz stats: temporary fix for pds collection inputs
The stats filters were crashing on input partitioned dataset
collections because it was producing partitioned datasets nested with a
multiblock in the output model. To fix it, the partitioned dataset is
replaced by a multiblock in this scenario.

This is a temporary fix. Stat filters should be able to handle
partitioned datasets correctly, and not compute on stat per partition.

Fixes #20854