vtkSMRenderViewProxy: select GeometryRepresentation for composite data
Composite datasets may not have leaf datasets at the time of their
creation, hence knowing whether there is a suitable representation for
them in a RenderView is not known at the time the representation type
is queried. This can lead to all kinds of problems, such as 21381,
21293, and 21041.

This commit implements a mostly good workaround where a
GeometryRepresentation will be created for composite datasets. This
may be incorrect for when the composite dataset will be populated with
vtkTable leaves, for instance, but that could happen in previous
versions of ParaView if you grouped geometry datasets with the Group
Datasets filter and later changed the inputs of the filter to only
vtkTables (a rare case). Further work may be needed to resolve that
issue. The worst that can happen with this change is that a
representation that doesn't show anything for vtkTables will be
created in a RenderView.