Disable the "Solid Color" choice for the "Slice" and "Volume" reprs
The "Slice" representation for images throws an error and keeps
displaying the active scalar if "Solid Color" is chosen to color-by.
The "Volume" representation shows a blank visualization with no scalar
field.

Track the representation type in the pqDisplayColorWidget and disable
the "Solid Color" entry if the "NoSolidColor" hint is present for the
current representation and representation type.