Empty the "Favorites" list in pqFileDialog by default
Also fix a sizing issue with the "Locations" list, and move the
static variable controlling whether "Examples" is added to the list,
since it only applies to the Location list now.

Add "Downloads" to the Windows special directories, and update the
function used to retrieve them to Vista+ standards.

Make sure "Examples" shows as invalid in the Locations list
if it doesn't exist.