NodeEditor: add saving layout capabilities
When saving and reloading a state file with a complex pipeline, if one has took time to setup a nice layout in the node editor then there work would be lost.
Now a file is saved alongside the statefile that saves the pipeline layout.

Add new signals to pqApplicationCore in order to catch the filename of the saved/loaded state file.