add(GmshIO): support for msh2 format on write
The GmshIO plugin now supports writing v2.2 gmsh files when using the
`*.msh2` extension for the file name.

The GmshIO plugin uses the gmsh c api as a backend to write .msh files
from `vtkUnstructuredGrid`s. By default, these files are written in
gmsh's 4.1 file format. However, if given a file with a different
extension, the behavior of the underlying c api is modified to
accomodate that extension. As such, by simply allowing for files with
the `*.msh2` extension, one enables users to write gmsh files in the 2.2
format.

Default behavior if no extension is provided is still to use the `*.msh`
extension and thus the 4.4 format.