Wrapping/Python: remove import of the static module
This was added in 8b3c13ec9d (Import the static Python module loader,
add __all__, 2022-04-29) (via !5673) to help out the `.pyi` generation
in a static build ensure that the builtin modules are loaded so that the
classes can be queried. However, this causes problems in static builds
which link to ParaView itself and load the Python modules, notably in
the Protobuf type registry calling `abort()` when duplicate message
types are seen. This is because the tool which links to ParaView
directly loads it and then when it loads up Python code (in this
instance, via Catalyst 1's Python pipeline support), the Python static
module is imported and causes a crash.

The `.pyi` generation now knows how to do this on its own and such
support does not need to be done globally.