Link widget colors to the color palette.
Linking colors allows the user to change widget colors in the
ParaView settings for the color palette, and makes sure that
when the palette swaps the foreground/background colors like with
the White Background palette, the widget edges and features are
still visible. Widgets link to the "Foreground", "Interactive Widget",
and "Widget" colors.

Adds a "Widget" color to the palette, as a third color that widgets
can use to distinguish available handles and handles that are
actively being manipulated. Not all widgets use all three colors (yet).

Update VTK for f32773b9ec, which includes the widget color setters.