Hide the Window Context Help item.
Complete removal of context help. It is used rarely, and adds visual clutter,
so it is replaced by tooltips whenever possible.

On Windows, this appears as a "?" in the window Title Bar.
On Linux, this appears as a menu item in the Title Bar menu

This is a redo of an out-of-date MR by @stan.hammon