Add a setting to add a location to the file dialog from an env-var
A new setting in the Misc section allows the user to specify a list
of labels and environment variables. Each of the env-var values
is looked up, and if it exists, the path it contains is added to
the Locations list in the File dialog. If the path doesn't exist,
the warning icon is used. If the env-var is not set or is empty,
the entry is not added to the Locations list.

Add handling of `ShowComponentLabels` hint to
`pqStringVectorPropertyWidget` so the settings widget has sensible
labels.