Implement callback for "shownavigationpanel" action
Fix warning about unbound "shownavigationpanel" action.
Provide a basic implementation but do not add it to any JSON file as it is
not a common feature. (i.e. it shows the position of the controller in real
time using a vtkVRPanel).
Users can always use this action by editing their JSON file and mapping it
to a click action, e.g.:

  {
    "inputs": {
      "click": {
        "output": "shownavigationpanel"
      }
    },
    "path": "/user/hand/left/input/b"
  }