Allow individual properties to save and reset defaults in settings
Using a context menu available on the label or the main widget if
there is no label, show a context menu to save the current value
for a property as the default, or reset it to the application
default value, mirroring the buttons that perform those actions for
the entire section of properties.

Add an XML test.