Make shortcuts active in the entire window by default.
If shortcuts are active only in the widget and children which is
their parent, they can only be activated if that widget has focus.
Users expect shortcuts to be active when the render window doesn't
have focus, like after a filter is created or clicked on in the
pipeline window.

Modify the point picking helper to check if the mouse is inside
the render window before performing a pick - other shortcuts that
depend on the mouse position should follow this pattern.