Fix vtkmContour test failure
VTK-m's contour filter has re-enabled the flying edges algorithm.
The result generated by that is slightly different, where the generated
scalar field has small variances, causing the color mapping to be different
enough for the test to fail. Here we re-scale the color mapping range to be
larger so that the small variance does not affect the color mapping.