Add save homepage action
Add new action to the help window to save the current page as home page. The homepage url is saved in the settings.
All help window related classes are moved from the pqWidgets module to the pqComponents one in order to access to pqSettings (pqCore can't be a dependency of pqWidgets).