Fix a small issue in array information
When displaying a string array in information panel,
the strings are not separated by a colon as expected.

It was caused by a missing counter incrementation. Indeed, the colon
is added only if the counter is increased after the first iteration.

This incrementation was missing as the loop had been previously reworked
into a more modern (for var : list) approach but the needed incrementation was overlooked.

This commit fixes this by incrementing the counter at the end of the loop.

This commit also fixes a small warning related to types.