Selection Editor: add color array
To be able to visualize a selection with different colors, we can now
set in the selection editor for each selection a specific color.

Each color specify will be evaluate in the vtkAppendSelection to
generate a new data array during the combineActiveSelection.

This new array will be added during the vtkExtractSelection.