Add Bivariate Representations Plugins
This plugin brings a new representation to visualize two field data arrays at the same time.
It uses Perlin noise to visualize the values of the 2nd array: the noise is stronger where the values are high.
Only scalar point data can be chosen as 2nd array.
This representation can be used to visualize 2D and 3D surfacic data.