vtkSimulationToPrismFilter: Fix interpolation and Split in 2 filters
vtkSimulationToPrismFilter has been split in 2 filters:
1) vtkSimulationPointCloudFilter:
   1) cell interpolation or point interpolation based on the attribute type
      This is done using vtkPointDataToCellData, vtkCellDataToPointData
   2) creates dummy points
   3) creates as many vertices as the number tuples of the attribute type
2) vtkSimulationToPrismFilter:
   1) Creates output points using vtkMergeVectorComponents
   2) Copies vertices/cell data/point data from vtkSimulationPointCloudFilter

It has been done to simplify the code by using well-tested filters, and also
improving the performance by calling filters only when they are needed.
Previously all the filters would be called when you change an array, but
now only the vtkMergeVectorComponents in vtkSimulationToPrismFilter.
will be called. This is done for better performance when dealing with large
simulation datasets.