vtkSMFieldDataDomain: Fix default_values selection
vtkSMFieldDataDomain::ComputeDefaultValues was faulty in its logic
interpreting enum values as indices.

This was working by chance on classic point/cell/field dataset
but not working at all with other datasets.

This fixes it by considering the default_values as an enum value,
eg vtkDataObject::FIELD or vtkDataObject::POINT as everywhere else
in vtkSMFieldDataDomain.