Adds NUMSERVERS single valued argument
Adds NUMERSERVERS single valued argument in the _paraview_add_test function.
It is not used but mandatory to avoid being filtered out from ARGN in the caller.
If it was not filtered out and not present as an argument, then it would be considered as part of TEST_SCRIPT multi-values arguments.