Bring vtkMultiDimensionalArray to DSP plugin
Add the vtkMultiDimensionalArray to DSP plugin. For now, this implicit array allows to represent 3D arrays
and provides "2D views" on it. The motivation is to use in the DSP plugin in order to avoid relying on
multiblock datasets to represent temporal data on each point of a mesh (output of the "Plot Over Time" filter),
that are not scalable in a multi-process environment.