Make pqCustomViewpointsToolbar widget modulable
pqCustomViewpointsToolbar widget is now associate with a "controller" that
enable the toolbar to be used for both desktop (classic) paraview and
paraview VR/XR. The pqCustomViewpointsController is a base class that
contains the same interfaces has the old pqCustomViewpointsToolbar, which
are called by the toolbar when needed. pqCustomViewpointsDefaultController
contains the part of the pqCustomViewpointsToolbar old code, which controls
the desktop view.