Add a camera widget view link
The proxy link is augmented with a new property to know how to handle
exceptions. They can be a black list (as before, default) or a white
list. Camera widget view link needs no properties linked, so we just set
it to white list with no exception.

Add link beside interactive view link in camera link and link manager.