Link update views only when asked to
* When updating a RenderView property, `UpdateVTKObject` is called
and we finally updates all views.
Moving camera leads to another code path. So we observe the render
`EndEvent` to trigger a rendering of the other views.
Unfortunately, when Camera properties are excluded from the link
AND if a CameraLink is also created, we fall into a view freeze
issue (see !22193)