Saving screenshots:
pvpython was invalidating pointers used in the threaded callback queue
if the program wanted to exit before the last task was terminated. It
was doing so because the callback queue was a static instance of
'vtkProcessModule`. We actually want it destroyed inside the destructor
of `vtkProcessModule`, so the queue is not static anymore.