vtkGeometryRepresentation: Fix PopulateBlockAttributes
For PDC, we used to convert assembly selectors to hierarchy selectors.
if there were blocks with the same node name but different path name,
since in hierarchy the path name is not present, there were duplicate node names = path names.
Duplicate path names can't be resolved, and that's why there was a bug.
If the hierarchy of such PDC with duplicate node names, is accessed, the bug is still present.
Make sure that the path name is unique, but the node name doesn't need to be unique.

# paraview/paraview/-/issues/22216