Update Palette Background baselines
* Looks like we had false positive palette test. Some intermediate
screenshots used to occurs before the actual update, and baselines
reflected this wrong result.
This was introduce in ParaView MR !5287.

A VTK update on Background handling (see vtk
ffd07e2e3877f465830004cac64a9676ccbd14da) changes the rendering
stack for background, highlighting this issue.

So the fix is to update our baselines.