pqFileDialog: Remove existence check when selecting multiple files
When files are selected 1 existence check (plus another 1 even more useless) was performed
to enable or not the OK button.

This introduces an overhead just for browsing files, which is not necessary
since when you press OK, a.k.a. when it matters, the existence check will be performed.

paraview/paraview/-/issues/22230