vtkGeometryRepresentation: Fix PopulateBlockAttributes (2)
For PDC, we used to convert assembly selectors for properties to hierarchy selectors.
If there were blocks with the same node name but different path name,
since in hierarchy the path name is not present, there were duplicate node names = path names.
Duplicate path names can't be resolved, and that's why there was a bug.

Also support pdc for subclass representations that generate a pdc