Improve MultiDimensionBrowser for distributed environment
* The Index parameter was set for the hidden dimension on each rank.
With N timesteps and R ranks, this leads to NxR rows for a given index.
So final table depends on number of ranks. This does not make sense.
* We aim to have N rows in output table. So only one rank should contribute to output,
with a deterministic heuristic.
* We now use a global id to determine wich index of which rank should be
the output.
This global id array can be provided externally, or we implicitly
attribute a local offset depending on the rank number and size of the previous ranks
* Then only the rank containing Index in the global id array is used