Handle having a semicolon or hash in directory or filenames.
Semicolon ";" is used as the separator between the path and key-value pairs for
additional data needed by recent file entries, and hash "#" is used to
designate a server name, so encode those characters for serialization. Don't
use the more general QUrl::fromPercentEncoding(), because that prevents
using "%" in filenames at all.