help: move definition of StringWriter to the end
For an unknown reason, having this proxy at the start of the file
resulted in it being the only writer listed in the online
documentation. Moving to the end has the effect of letting other
writers appear in the documentation, but not the StringWriter. This
isn't a perfect fix, but improves the reader documentation.