Search for libdice.so also in the directory that contains libnvindex.so.
When the ParaView plugin has its RPATH set to the directory that contains
libnvindex.so then IndeX will also be able to load libdice.so that is located in
the same directory.

However, when RUNPATH is used instead of RPATH, which may happen depending on
the build infrastructure, this won't work anymore because RUNPATH is not
transitive.

IIndex_loader_configuration allows setting the path used for loading libdice.so:
It will try the directory containing libnvindex.so before falling back to the
default search mechanism (to accommodate LD_LIBRARY_PATH).

See also paraview-superbuild#236.