Fix handling of (signed) "char" and "short" data for colormaps and on aarch64.
"char" (VTK_CHAR) is signed on x86-64, but unsigned on aarch64 and ppc64le. This
is now handled correctly when rendering structured volumes, but will output a
warning, recommending to use "signed char" (which is now supported) or "unsigned
char" instead. This is similar to what the "Volume" representation does (which
won't render "char" data at all).

The colormap domain for "(signed) char" and "short" (but not for "int", unsigned
integer and floating point types) was previously normalized to [0.0, 1.0] but is
now normalized to [-1.0, 1.0], to be consistent with other ParaView
representations (e.g. "Surface" and "Volume").