fides: disable create shared points property by default
For image data that doesn't have ghost points, Fides will create shared
points to close gaps between blocks. However there are cases where its
trying to create shared points when it shouldn't - 1) when there are
ghost points, or 2) when it's not image data. Until vtk/fides#19 is
fixed (and changes make their way to VTK and ParaView), we'll set this
property to be off by default.