vtkSpreadSheetView: Ensure that a block is always present during update
vtkSpreadSheetView::Update() adds a block to the cache if none exists.
This is needed to ensure that if a progress event is attempted to be
handled, because of a WrongTagEvent during another Receive call message
from the client's controller in any view that is present, the progress can
query some block from the cache instead of trying to fetch it from the
server which will result in a deadlock. After finishing the Update(), a
StillRender/InteractiveRender will be called which will ensure that the
cache is cleared and the correct block is fetched from the server.