feat(ConstantShading): Add constant shading per object.
Can now control if a specific object in the pipeline is shaded
using the scene lighting or not. It is equivalent to disabling the
Light Kit but can be done per object, not only for the
whole view. A checkbox has been added in the `Lighting`
section of the property panel for the representations that
supports it. There are currently some bugs that originate from
specific shaders implementation (Surface LIC for instance),
but those are also present when using the Light Kit so they
won't be addressed here.