Add new fullscreen mode
Add new fullscreen mode showing only the current active view in fullscreen.
The current fullscreen mode set the tabbed multi-view widget fullscreen, i.e. all views are maximized, including
separators and decoration. The new mode is intended to be closer to a "classical" fullscreen mode.
This is in part motivated by the need of some hardware (like the zSpace Inspire) to have a strict fullscreen mode
in order to display the rendered scene correctly.