Add vtkAxisAlignedCutter filter
Add vtkAxisAlignedCutter filter to VTKExtension and expose it in ParaView.
This filter only accepts Axis-Aligned implicit function to slice the input data.
This allows to preserve the input data type. In other words, it reduces the input
data dimension by 1.
This filter supports for now HTGs and OverlappingAMRs. It only suports Axis-Aligned
Planes. For HTG inputs, this filter also support multiple slices. In that case,
the output is a PartitionedDataSetCollection containing all the slices.