Handle fullscreen state on zSpace devices
Notify the zSpace SDK when we activate or deactivate "ActiveViewFullScreen" mode.
This allows to enable the autostereolens of the zSpace Inspire harware when stereo can be displayed (only in fullscreen)
and disable it otherwise. This is useful notably to display UI elements of ParaView or other windows without the autostereolens
activated (which distort non-stereo elements).