Fix deprecated positional option for '.pvx' file
Since CLI11 2.4.0, the `'.pvx' file` option is not recognized as a valid
option and results in the following runtime error:

    terminate called after throwing an instance of 'CLI::BadNameString'
      what():  Invalid positional Name: '.pvx' file
    Aborted (core dumped)

Hence, when ParaView is compiled with a new CLI11 version, starting
`pvserver` or `pvrenderserver` always fails before the user even sees
the result of CLI parsing.

Since the option is deprecated, I don't see much value in actually
fixing it and only guarded it with CLI11 version macros, so the option
will not be available when a new CLI11 version is used.