Add --venv argument to all executables
Adds a new command-line argument that is the path of a virtual
environment that should be used by the Python interpreter in each
executable when Python is enabled. The virtual environment is one
produced by the virtualenv tool.

This feature makes it possible to directly invoke a package installed
in a virtual environment.