Undock MultiBlock Inspector in tests
The MultiBlock Inspector starts out docked on the right side of the
main window.  This squeezed the view on macOS to the point of changing
the zoom factor in the generated test images, causing test
failures. This is solved by undocking the MultiBlock Inspector as soon
as it is shown.