Update AMReXParticlesReader test and its baseline
In the past vertices used to be created for each point by the reader,
(see vtkAMReXParticlesReader line 241), if no cells are presents,
bounds are incorrect, therefore we need to recenter the glyphs to make them visible.

Also, the old glyph filter used to generate slightly different results, and that's
why we update the baseline.