Fix SelectionByArray and update its baseline
The default of slices has changed, that's why it was added.
The compressed arrays don't enable selections by cell data,
because only vtkIdTypeArray are supported, that's why we
expose the default and set it to 0.

The baselines needed to be updated because Surface with Edges now
are not as sharp as they used to be which makes them look thicker.
In some machines the angle is slightly different that's why 2 baselines.