vtkDataLabelRepresentation: hide point labels when input is not correct
Upon receiving REQUEST_RENDER with an invalid producer object, hide point label actor. This was already the case for cell labels before.
This could cause problems for example when loading a state file containing a vtkDataLabelRepresentation proxy, and update of the view was requested before the view input had been properly setup.