simple.py: Add ListProperties() function
This function can list the properties of either an existing proxy OR a
proxy creation function, e.g., paraview.simple.Sphere,
paraview.simple.Clip, etc. This function can be used to get a list of
properties *without* needing to first create a proxy.

The property list consists of property labels because:

- It matches what is seen in the GUI.
- It matches what is seen when the key sequence object.<TAB> is typed
  in the Python shell.

Co-authored-by: Christos Tsolakis <christos.tsolakis@kitware.com>