ParaView offers access to standard ApplicationData directory
* Qt provides way to get standard directory paths (see QStandardPaths)
This adds an API in ParaView to get the relevant AppLocalDataLocation.

This is for instance `.local/share/ParaView/ParaView` on linux.
The redundant `ParaView` directory is because we specify it for both
OrganizationName and an ApplicationName in the QApplication.

This will be useful to store some runtime files (like autosave)