InsituLive: improve live visu connect/disconnect mechanism
Previously, there was no button in the Catalyst toolbar to disconnect to
a catalyst server. User can only do that by using the `Files >
Disconnect` button which disconnect only the active server. If it
disconnects the wrong server, the simulation can hang indefinitely.

Now, there is a dedicate button `Catalyst > Disconnect` which make sure
to disconnect only the catalyst server. If a user tries to disconnect by
using the previous method it will have a warning saying to use this new
button instead.