version: bump to 5.13
This fixes the version numbering ordering between `master` and
`release`.

This was missed for RC1 and planned for RC2, but missed there as well.
The superbuild now sees the discrepency and gets confused about the
mismatch due to RC2's tag being properly visible despite RC1 being
reachable. It would have happened for RC1 except for this Git bug:

    https://lore.kernel.org/git/ZNffWAgldUZdpQcr@farprobe/T/#u